#55 - Tests refactoring
