+++ src/test/java/dev/iakunin/codexiabot/codexia/cron/CodexiaParserIntegrationTest.java
@@ -3,28 +3,22 @@ package dev.iakunin.codexiabot.codexia.cron;
 import com.github.database.rider.core.api.dataset.DataSet;
 import com.github.database.rider.core.api.dataset.ExpectedDataSet;
 import dev.iakunin.codexiabot.AbstractIntegrationTest;
-import dev.iakunin.codexiabot.CodexiaBotApplication;
+import dev.iakunin.codexiabot.config.GithubConfig;
 import dev.iakunin.codexiabot.util.WireMockServer;
 import dev.iakunin.codexiabot.util.wiremock.Stub;
-import lombok.SneakyThrows;
 import org.cactoos.io.ResourceOf;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.Test;
-import org.kohsuke.github.GitHub;
-import org.kohsuke.github.GitHubBuilder;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.test.util.TestPropertyValues;
 import org.springframework.context.ApplicationContextInitializer;
 import org.springframework.context.ConfigurableApplicationContext;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.context.annotation.Import;
 import org.springframework.test.context.ContextConfiguration;
 
 @SpringBootTest(classes = {
     AbstractIntegrationTest.TestConfig.class,
-    CodexiaParserIntegrationTest.TestConfig.class,
+    GithubConfig.class,
 })
 @ContextConfiguration(initializers = CodexiaParserIntegrationTest.Initializer.class)
 public class CodexiaParserIntegrationTest extends AbstractIntegrationTest {
@@ -121,19 +115,6 @@ public class CodexiaParserIntegrationTest extends AbstractIntegrationTest {
             TestPropertyValues.of(
                 "app.codexia.base-url=" + WireMockServer.getInstance().baseUrl() + "/codexia"
             ).applyTo(applicationContext.getEnvironment());
-        }
-    }
-
-    @Configuration
-    @Import(CodexiaBotApplication.class)
-    static class TestConfig {
-        @SneakyThrows
-        @Bean
-        public GitHub gitHub() {
-            return new GitHubBuilder()
-                .withEndpoint(
-                    WireMockServer.getInstance().baseUrl() + "/github"
-                ).build();
         }
     }
 }
