+++ src/test/java/dev/iakunin/codexiabot/github/GithubModuleCreateRepoIntegrationTest.java
@@ -3,35 +3,29 @@ package dev.iakunin.codexiabot.github;
 import com.github.database.rider.core.api.dataset.DataSet;
 import com.github.database.rider.core.api.dataset.ExpectedDataSet;
 import dev.iakunin.codexiabot.AbstractIntegrationTest;
-import dev.iakunin.codexiabot.CodexiaBotApplication;
+import dev.iakunin.codexiabot.config.GithubConfig;
 import dev.iakunin.codexiabot.github.GithubModule.RepoNotFoundException;
 import dev.iakunin.codexiabot.util.WireMockServer;
 import dev.iakunin.codexiabot.util.wiremock.Response;
 import dev.iakunin.codexiabot.util.wiremock.Stub;
 import java.io.IOException;
-import lombok.SneakyThrows;
 import org.cactoos.io.ResourceOf;
 import org.cactoos.text.FormattedText;
 import org.junit.jupiter.api.AfterEach;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 import org.junit.jupiter.api.Test;
-import org.kohsuke.github.GitHub;
-import org.kohsuke.github.GitHubBuilder;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.test.util.TestPropertyValues;
 import org.springframework.context.ApplicationContextInitializer;
 import org.springframework.context.ConfigurableApplicationContext;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.context.annotation.Import;
 import org.springframework.http.HttpStatus;
 import org.springframework.test.context.ContextConfiguration;
 
 @SpringBootTest(classes = {
     AbstractIntegrationTest.TestConfig.class,
-    GithubModuleCreateRepoIntegrationTest.TestConfig.class,
+    GithubConfig.class,
 })
 @ContextConfiguration(initializers = GithubModuleCreateRepoIntegrationTest.Initializer.class)
 public class GithubModuleCreateRepoIntegrationTest extends AbstractIntegrationTest {
@@ -173,19 +167,6 @@ public class GithubModuleCreateRepoIntegrationTest extends AbstractIntegrationTe
             TestPropertyValues.of(
                 "app.codexia.base-url=" + WireMockServer.getInstance().baseUrl() + "/codexia"
             ).applyTo(applicationContext.getEnvironment());
-        }
-    }
-
-    @Configuration
-    @Import(CodexiaBotApplication.class)
-    static class TestConfig {
-        @SneakyThrows
-        @Bean
-        public GitHub gitHub() {
-            return new GitHubBuilder()
-                .withEndpoint(
-                    WireMockServer.getInstance().baseUrl() + "/github"
-                ).build();
         }
     }
 }
